ensure one path for dealing with sentry
