+++ cmd/ssh_terminal.go
@@ -15,7 +15,7 @@ import (
 	"github.com/superfly/flyctl/api"
 	"github.com/superfly/flyctl/cmdctx"
 	"github.com/superfly/flyctl/helpers"
-	"github.com/superfly/flyctl/internal/flyerr"
+	"github.com/superfly/flyctl/internal/sentry"
 	"github.com/superfly/flyctl/pkg/agent"
 	"github.com/superfly/flyctl/pkg/ssh"
 	"github.com/superfly/flyctl/terminal"
@@ -70,9 +70,9 @@ func runSSHConsole(cc *cmdctx.CmdContext) error {
 			return
 		}
 
-		flyerr.CaptureException(err,
-			flyerr.WithTag("feature", "ssh-console"),
-			flyerr.WithContexts(map[string]interface{}{
+		sentry.CaptureException(err,
+			sentry.WithTag("feature", "ssh-console"),
+			sentry.WithContexts(map[string]interface{}{
 				"app": map[string]interface{}{
 					"name": app.Name,
 				},
